Remove search service
