--- deployments/local-minio/configs/mis_config.yml
@@ -1,11 +1,0 @@
-aws:
-  s3:
-    endpointUrl: http://minio:9000
-    region: local
-    bucket:
-      quarantine: ingest-quarantine
-solr:
-  host: solr
-  port: 8983
-endpointUrl:
-  retrieve: http://localhost:9041/mis/product/
