+++ deployments/local-minio/configs/store_config.yml
@@ -1,0 +1,2 @@
+endpointUrl:
+  retrieve: http://localhost:9041/mis/product/
