--- deployments/local-minio/configs/transform_config.yml
@@ -1,2 +1,0 @@
-endpointUrl:
-  transform: http://transformation:9090/transform/
