+++ settings.gradle
@@ -5,5 +5,5 @@
  * https://www.gnu.org/licenses/lgpl-3.0.html
  */
 /* Build Script */
-rootProject.name = "content-discovery-retrieval"
+rootProject.name = "store"
 include "store"
