+++ store/build.gradle
@@ -20,13 +20,12 @@ dependencies {
     annotationProcessor "org.projectlombok:lombok:1.18.8"
     compileOnly "org.projectlombok:lombok:1.18.8"
     implementation "com.amazonaws:aws-java-sdk-s3:1.11.609"
-    implementation "com.connexta.search:index-api-rest-spring-stubs:0.1.0-SNAPSHOT" // TODO replace with released version
     implementation "javax.inject:javax.inject:1"
     implementation "io.springfox:springfox-swagger2:2.9.2"
     implementation "io.springfox:springfox-swagger-ui:2.9.2"
     implementation "org.springframework.boot:spring-boot-starter-actuator:2.1.7.RELEASE"
     implementation "org.springframework.boot:spring-boot-starter-web:2.1.7.RELEASE"
-    implementation "org.springframework.data:spring-data-solr:4.0.10.RELEASE"
+    implementation "commons-io:commons-io:2.6"
     testImplementation "com.github.npathai:hamcrest-optional:2.0.0"
     testImplementation "com.xebialabs.restito:restito:0.9.3"
     testImplementation "junit:junit:4.12"
@@ -45,7 +44,7 @@ bootJar {
 
 docker {
     files tasks.bootJar.outputs
-    name "cnxta/cdr-store:${project.version}"
+    name "cnxta/ion-store:${project.version}"
     buildArgs([JAR_FILE: "${tasks.bootJar.outputs.files.singleFile.name}"])
 }
