+++ store/src/main/java/com/connexta/store/adaptors/StorageAdaptor.java
@@ -6,7 +6,7 @@
  */
 package com.connexta.store.adaptors;
 
-import com.connexta.store.common.exceptions.StorageException;
+import com.connexta.store.common.exceptions.StoreException;
 import java.io.IOException;
 import java.io.InputStream;
 import javax.validation.constraints.Max;
@@ -20,7 +20,7 @@ public interface StorageAdaptor {
    * Stores the {@code file} in a blob store using the given {@code key}.
    *
    * @throws IOException if the file can't be read
-   * @throws StorageException if there is an error when attempting to store
+   * @throws StoreException if there is an error when attempting to store
    */
   void store(
       @NotNull @Min(1L) @Max(10737418240L) final Long fileSize,
@@ -28,16 +28,14 @@ public interface StorageAdaptor {
       @NotBlank final String fileName,
       @NotNull final InputStream inputStream,
       @NotBlank final String key)
-      throws StorageException;
+      throws StoreException;
 
   /**
    * Retrieves the data in a blob store using the given key.
    *
    * @param key the key used to reference the stored object
-   * @throws StorageException
+   * @throws StoreException
    */
   @NotNull
-  RetrieveResponse retrieve(@NotBlank final String key) throws StorageException;
-
-  boolean objectExists(@NotBlank String key);
+  RetrieveResponse retrieve(@NotBlank final String key) throws StoreException;
 }
