--- store/src/main/java/com/connexta/store/common/exceptions/SearchException.java
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2019 Connexta, LLC
- *
- * Released under the GNU Lesser General Public License version 3; see
- * https://www.gnu.org/licenses/lgpl-3.0.html
- */
-package com.connexta.store.common.exceptions;
-
-public class SearchException extends Exception {
-
-  public SearchException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
