+++ store/src/main/java/com/connexta/store/config/S3StorageConfiguration.java
@@ -11,8 +11,6 @@ import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.amazonaws.services.s3.transfer.TransferManager;
-import com.amazonaws.services.s3.transfer.TransferManagerBuilder;
 import com.connexta.store.adaptors.S3StorageAdaptor;
 import java.io.File;
 import java.io.IOException;
@@ -47,9 +45,7 @@ public class S3StorageConfiguration {
   public S3StorageAdaptor s3StorageAdaptor(
       @NotNull final AmazonS3 amazonS3,
       @Value("${aws.s3.bucket.quarantine}") @NotBlank final String bucket) {
-    TransferManager transferManager =
-        TransferManagerBuilder.standard().withS3Client(amazonS3).build();
-    return new S3StorageAdaptor(amazonS3, transferManager, bucket);
+    return new S3StorageAdaptor(amazonS3, bucket);
   }
 
   @Bean
