--- store/src/main/java/com/connexta/store/config/SearchServiceConfiguration.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2019 Connexta, LLC
- *
- * Released under the GNU Lesser General Public License version 3; see
- * https://www.gnu.org/licenses/lgpl-3.0.html
- */
-package com.connexta.store.config;
-
-import com.connexta.store.repositories.IndexedMetadataRepository;
-import com.connexta.store.services.api.SearchService;
-import com.connexta.store.services.impl.SearchServiceImpl;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SearchServiceConfiguration {
-
-  @Bean
-  public SearchService searchService(
-      @NotNull IndexedMetadataRepository indexedMetadataRepository,
-      @NotBlank @Value("${endpointUrl.retrieve}") String retrieveEndpoint) {
-    return new SearchServiceImpl(indexedMetadataRepository, retrieveEndpoint);
-  }
-}
