+++ store/src/main/java/com/connexta/store/config/StorageManagerConfiguration.java
@@ -7,10 +7,8 @@
 package com.connexta.store.config;
 
 import com.connexta.store.adaptors.StorageAdaptor;
-import com.connexta.store.common.IndexManager;
-import com.connexta.store.common.ProductStorageManager;
-import com.connexta.store.models.IndexedProductMetadata;
-import com.connexta.store.services.api.Dao;
+import com.connexta.store.service.api.StoreService;
+import com.connexta.store.service.impl.StoreServiceImpl;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Value;
@@ -19,16 +17,11 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class StorageManagerConfiguration {
-  @Bean
-  public ProductStorageManager productStorageManager(
-      @NotBlank @Value("${endpointUrl.retrieve}") String retrieveEndpoint,
-      @NotNull StorageAdaptor storageAdapter) {
-    return new ProductStorageManager(retrieveEndpoint, storageAdapter);
-  }
 
   @Bean
-  public IndexManager metadataStorageManager(
-      @NotNull Dao<IndexedProductMetadata, String> cstDao, @NotNull StorageAdaptor storageAdaptor) {
-    return new IndexManager(cstDao, storageAdaptor);
+  public StoreService productStorageManager(
+      @NotBlank @Value("${endpointUrl.retrieve}") String retrieveEndpoint,
+      @NotNull StorageAdaptor storageAdapter) {
+    return new StoreServiceImpl(retrieveEndpoint, storageAdapter);
   }
 }
