--- store/src/main/java/com/connexta/store/models/Product.java
@@ -1,14 +1,0 @@
-/*
- * Copyright (c) 2019 Connexta, LLC
- *
- * Released under the GNU Lesser General Public License version 3; see
- * https://www.gnu.org/licenses/lgpl-3.0.html
- */
-package com.connexta.store.models;
-
-public class Product implements Storable<String> {
-  @Override
-  public String getId() {
-    return null;
-  }
-}
