--- store/src/main/java/com/connexta/store/models/Storable.java
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2019 Connexta, LLC
- *
- * Released under the GNU Lesser General Public License version 3; see
- * https://www.gnu.org/licenses/lgpl-3.0.html
- */
-package com.connexta.store.models;
-
-public interface Storable<E> {
-
-  E getId();
-}
