+++ store/src/main/java/com/connexta/store/service/impl/StoreServiceImpl.java
@@ -4,11 +4,12 @@
  * Released under the GNU Lesser General Public License version 3; see
  * https://www.gnu.org/licenses/lgpl-3.0.html
  */
-package com.connexta.store.common;
+package com.connexta.store.service.impl;
 
 import com.connexta.store.adaptors.RetrieveResponse;
 import com.connexta.store.adaptors.StorageAdaptor;
-import com.connexta.store.common.exceptions.StorageException;
+import com.connexta.store.common.exceptions.StoreException;
+import com.connexta.store.service.api.StoreService;
 import java.io.InputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -20,35 +21,34 @@ import javax.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class ProductStorageManager {
+public class StoreServiceImpl implements StoreService {
 
   @NotBlank private final String retrieveEndpoint;
   @NotNull private final StorageAdaptor storageAdaptor;
 
-  public ProductStorageManager(
+  public StoreServiceImpl(
       @NotBlank final String retrieveEndpoint, @NotNull final StorageAdaptor storageAdaptor) {
     this.retrieveEndpoint = retrieveEndpoint;
     this.storageAdaptor = storageAdaptor;
   }
 
-  public URI storeProduct(
-      @NotNull @Min(1L) @Max(10737418240L) final Long fileSize,
-      @NotBlank final String mediaType,
-      @NotBlank final String fileName,
-      @NotNull final InputStream inputStream)
-      throws StorageException, URISyntaxException {
+  @Override
+  public @NotNull URI createProduct(
+      @NotNull @Min(1L) @Max(10737418240L) Long fileSize,
+      @NotBlank String mediaType,
+      @NotBlank String fileName,
+      @NotNull InputStream inputStream)
+      throws StoreException, URISyntaxException {
     // TODO: Validate Accept-Version
     final String key = UUID.randomUUID().toString().replace("-", "");
 
     storageAdaptor.store(fileSize, mediaType, fileName, inputStream, key);
+
     return new URI(retrieveEndpoint + key);
   }
 
-  /**
-   * The caller is responsible for closing the {@link java.io.InputStream} in the returned {@link
-   * RetrieveResponse}.
-   */
-  public RetrieveResponse retrieveProduct(@NotBlank String id) throws StorageException {
+  @Override
+  public @NotNull RetrieveResponse retrieveProduct(@NotBlank String id) throws StoreException {
     return storageAdaptor.retrieve(id);
   }
 }
